% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_var.R
\name{coef_var}
\alias{coef_var}
\title{Calculate the coefficient of variation}
\usage{
coef_var(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
Calculate the coefficient of variation
}
\examples{
x <- rnorm(100)
coef_var(x)
}
