% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Value mismatching}
\usage{
x \%notin\% table
}
\description{
%notin% is simply the inverse of %in%
It is more convenient and readable than
!x %in% y
}
\examples{
x <- c(4, 5, 11)
y <- 1:10
x \%notin\% y
}
