% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_accuracy.R
\name{rmse}
\alias{rmse}
\title{Prediction accuracy}
\usage{
rmse(pred, obs)
}
\arguments{
\item{pred}{a numeric vector of predictions}

\item{obs}{a numeric vector of observations}
}
\description{
Various measures of prediction accuracy
}
\details{
rmse: root mean square error
rrmse: relative root mean square error
mpe: mean percentage error
}
